Develop

+++ src/main/java/com/jhops10/canteen_manager/dto/product/ProductUpdateDTO.java
@@ -2,9 +2,11 @@ package com.jhops10.canteen_manager.dto.product;
 
 import com.jhops10.canteen_manager.model.Product;
 
+import java.math.BigDecimal;
+
 public record ProductUpdateDTO(
         String productName,
-        Double unitValue
+        BigDecimal unitValue
 ) {
     public void applyUpdateTo(Product entity) {
         if (productName != null) entity.setProductName(productName);
